6.Size 2nd commit

+++ 0x00-hello_world/6-size.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 /**
- * main - Prints all type sizes
+ * main - Prints some type sizes
  *
  * Description - This program will print the size of various types on this
  * computer.
@@ -17,22 +17,22 @@
 int main(void)
 {
 
-	char cht; /* Character */
-	int itg;  /* Integer */
-	int litg;  /* Long integer */
-	int llitg; /* Long Long integer */
-	float flt; /* Float */
+	char cht;
+	int itg;
+	long litg;
+	long long llitg;
+	float flt;
 
 
-	printf("Size of a char: %d byte(s)\n", sizeof(cht));
+	printf("Size of a char: %ld byte(s)\n", sizeof(cht));
 
-	printf("Size of an int: %d byre(s)\n", sizeof(itg));
+	printf("Size of an int: %ld byre(s)\n", sizeof(itg));
 
-	printf("Size of a long int: %d byte(s)\n", sizeof(litg));
+	printf("Size of a long int: %ld byte(s)\n", sizeof(litg));
 
-	printf("Size of a long long int: %d byte(s)\n", sizeof(llitg));
+	printf("Size of a long long int: %ld byte(s)\n", sizeof(llitg));
 
-	printf("Size of a float: %d byte(s)\n", sizeof(flt));
+	printf("Size of a float: %ld byte(s)\n", sizeof(flt));
 
 
 	return (0);
